Fix DE-MCMC log likelihood function and data
log.dens.like function takes an extra parameter (par.names) so we need
to adjust it to fill this in automatically. Additionally the data object
is stored as a list of data.frames, and to use in the IS2 we need one
large data.frame, so we just rbind them with a do.call.